requirements.txt
streamlit
pandas
python-docx
pypandoc
PyPDF2
openpyxl
